# content/swagger-doc.yaml
openapi: 3.0.0
info:
  title: Articles API
  version: 1.0.0
  description: API documentation for managing articles
servers:
  - url: http://localhost:9000
paths:
  /api/articles:
    get:
      summary: Retrieve a list of articles
      responses:
        200:
          description: A list of articles
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: string
                        name:
                          type: string
                        description:
                          type: string
                          nullable: true
                        price:
                          type: number
                        model:
                          type: string
                          nullable: true
    post:
      summary: Create a new article
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                id:
                  type: string
                name:
                  type: string
                description:
                  type: string
                  nullable: true
                price:
                  type: number
                model:
                  type: string
                  nullable: true
      responses:
        201:
          description: Article created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      id:
                        type: string
                      name:
                        type: string
                      description:
                        type: string
                        nullable: true
                      price:
                        type: number
                      model:
                        type: string
                        nullable: true

  /api/articles/{id}:
    get:
      summary: Retrieve an article by ID
      parameters:
        - in: path
          name: id
          required: true
          description: The ID of the article to retrieve
          schema:
            type: string
      responses:
        200:
          description: An article object
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      id:
                        type: string
                      name:
                        type: string
                      description:
                        type: string
                        nullable: true
                      price:
                        type: number
                      model:
                        type: string
                        nullable: true
        404:
          description: Article not found

    put:
      summary: Update an existing article
      parameters:
        - in: path
          name: id
          required: true
          description: The ID of the article to update
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                description:
                  type: string
                  nullable: true
                model:
                  type: string
                  nullable: true
      responses:
        200:
          description: Article updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      id:
                        type: string
                      name:
                        type: string
                      description:
                        type: string
                        nullable: true
                      price:
                        type: number
                      model:
                        type: string
                        nullable: true

    delete:
      summary: Delete an article
      parameters:
        - in: path
          name: id
          required: true
          description: The ID of the article to delete
          schema:
            type: string
      responses:
        204:
          description: Article deleted successfully